prepends docker-eula-java to the entrypoint of walle
* if the underlying distribution is tagged with oracle

NOTE

* sed/verification pattern looks quite equal: extraction?

refs #3

+++ .travis/build.sh
@@ -35,12 +35,20 @@ docker tag -f "$REPOSITORY/java-$ARCH:$JAVA" "java:8-jdk"
 
 case "$CUSTOM" in
     walle* )
+        if [[ "$JAVA" =~ .*-oracle$ ]]; then sed -i -r '/ENTRYPOINT/ s!/usr/bin/supervisord!docker-eula-java", "/usr/bin/supervisord!g' "contrib/walle/Dockerfile"; fi
+
         docker build -t "$TAG:$TAGSPECIFIER"      \
                      --build-arg  SCRIPT="${SCRIPT#*@}"  \
                      --build-arg PACKAGE="${SCRIPT%%@*}" \
                      "contrib/walle"
 
-        docker run --rm "$TAG:$TAGSPECIFIER" -v
+        if [[ "$JAVA" =~ .*-oracle$ ]]; then
+             if docker run --rm -e eula-java=accept "$TAG:$TAGSPECIFIER" -v; then             true; fi
+             if docker run --rm -e eula-java=       "$TAG:$TAGSPECIFIER" -v; then false; else true; fi
+        else
+                docker run --rm                     "$TAG:$TAGSPECIFIER" -v
+        fi
+
         ;;
     * )
         if [[ "$JAVA" =~ .*-oracle$ ]]; then sed -i -r '/ENTRYPOINT/ s!/bin/tini!docker-eula-java", "/bin/tini!g' "$PROJECT/Dockerfile"; fi
